feat(lspconfig): add ltex-ls support.
NOTE: ltex-ls uses non-standard code actions (i.e. add-to-dictionaries,
hide-false-positives), which are not planned to be supported by
nvim-lspconfig. The result is that ltex-ls are very likely to have a
very high false positive rate to mark words as mispelled (which is
supposed to be fixed by using the above code actions) and you can't fix
them easily. There's workaround by using plugin: `ltex_extra.nvim` or
`ltex-ls.nvim` to configure this lsp instead of nvim-lspconfig.
Migrating to those plugins is to be considered.