feat(quarto): add quarto support.
Currently quarto is basically treated as a "clone" of rmarkdown.
Features of quarto such as it's better integration with other languages
like python, bash, julia, etc. are not fully exploited. Only its
integration with R is configured.

TODO: watch `quarto-dev/quarto-nvim`'s development. When it's get better
development maturity, will try to use that package. In this way the
integration with other languages can be better exploited (such as
auto-completions from multiple languges).

TODO: watch the maturity of treesitter's markdown parser. Currently it
is a very buggy package and I am hesitate to use it (which is a hard
dependency of `quarto-dev/quarto-nvim`). (Or ideally, there will be a
standalone quarto parser).