config(org): remap `org-cycle` to `<Localleader><Tab>`.
It seems that the default globally behavior of `<C-i>` (jump forward) is
hard coded in vim. (I haven't find a way to programatically called some
"function" that does the samething as `<C-i>`). Thus once `<C-i>` is
remmaped buffer locally, I haven't find a way to bind another key that
does the same thing as the default `<C-i>` (i.e. jump foward).

As a workaround, I have to bind another key to `org-cycle` because jump
forward is still useful.