config(yarepl): Make more REPL mappings filetype-specific.
This change modifies more keymappings to be specific to the REPL
associated with the current buffer's filetype. For example,
`<Plug>(REPLFocus)` is now `<Plug>(REPLFocus-radian)` for R files and
`<Plug>(REPLFocus-ipython)` for Python files. This ensures that actions
are always sent to the correct REPL instance.

Additionally, the automatic REPL mapping for markdown, rmd, and quarto
files has been removed, as it was overly broad.