refactor(snacks): rework config setup.
We deliberately avoid using the higher-order abstractions provided
by lazy.nvim, such as combining multiple `opts` functions to merge the
options table that is eventually passed to the `xxx.setup` function, or
using `keys` to define keybindings. The rationale behind this is to keep
the configuration independent of any plugin manager's API, ensuring that
migration to another package manager will be simpler. (If ever required
in the future, though there are no current plans.)